Change type of function

+++ src/Intersection.h
@@ -1,6 +1,6 @@
 #ifndef GEOMETRY_INTERSECTION_H
 #define GEOMETRY_INTERSECTION_H
 
-int Intersection(float x1, float y1, float x2, float y2, float r1, float r2);
+int Intersection(int x1, int y1, int x2, int y2, int r1, int r2);
 
 #endif // GEOMETRY_INTERSECTION_H
